去掉了顶点的颜色属性,uv插值用z修正,有角度时图片不再失真

+++ inc/camera.h
@@ -17,7 +17,7 @@ public:
 	}
 	~Camera() {}
 
-	float l, r, t, b, n, f; //frustum参数
+	double l, r, t, b, n, f; //frustum参数
 
 	Transform transform;
